fix(ingestion): ignore inert and foreign-content markup in QA tables
The tree walk that collects table rows descended into subtrees that are not
part of the table:

- A <template>'s content is parsed into the ordinary child list (the parser
  has no separate template-contents field), so its placeholder rows were read
  as rows of the enclosing table. A user-supplied HTML page lost no data but
  gained a fabricated Q&A pair, e.g. a template row became
  "问题：模板问题\t回答：模板答案".
- Foreign content reuses HTML tag names: an <svg><tr> or <math><tr> carries no
  table semantics but was matched by tag name alone.

Skip inert subtrees at the top of the walk, and match table tags through a
helper that requires the HTML namespace. Keep the walk's existing behaviour of
not descending into a row it has already collected.

The skip list stays narrow: elements whose content is rendered (textarea,
span, svg text, …) keep their text.

Also split the general table-extraction cases into their own file so
xlsx_qa_regression_test.go keeps its name and the diff shows a modification
rather than a delete plus an add, and add a benchmark so the cost of the
extraction can be measured with:

  bash build.sh --test -run '^$' -bench BenchmarkExtractQATable -benchmem \
    ./internal/ingestion/component/chunker/
